Refactor 55 use big decimal instead of double

+++ src/main/java/jewellery/inventory/dto/request/ProductPriceDiscountRequestDto.java
@@ -3,6 +3,7 @@ package jewellery.inventory.dto.request;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import jakarta.validation.constraints.PositiveOrZero;
+import java.math.BigDecimal;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
@@ -16,6 +17,6 @@ import java.util.UUID;
 @AllArgsConstructor
 public class ProductPriceDiscountRequestDto {
   @NotNull private UUID productId;
-  @NotNull @Positive private Double salePrice;
-  @NotNull @PositiveOrZero private Double discount;
+  @NotNull @Positive private BigDecimal salePrice;
+  @NotNull @PositiveOrZero private BigDecimal discount;
 }
